migrations

+++ migrations/versions/089c71ee9a92_coins_history.py
@@ -20,9 +20,9 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "coins_history",
-        sa.Column("id", sa.INTEGER, primary_key=True, autoincrement=True, nullable=False, unique=True),
-        sa.Column("user_id", sa.INTEGER, sa.ForeignKey("users.id")),
-        sa.Column("amount", sa.DECIMAL),
+        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True, nullable=False, unique=True),
+        sa.Column("user_id", sa.Integer, sa.ForeignKey("users.id")),
+        sa.Column("amount", sa.Float),
         sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.func.now()),
     )
